+++ src/main/java/wooteco/subway/domain/Section.java
@@ -1,0 +1,5 @@
+package wooteco.subway.domain;
+
+public class Section {
+
+}
