--- src/main/java/wooteco/subway/exception/SubwayException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class SubwayException extends RuntimeException {
-
-    public SubwayException(String message) {
-        super(message);
-    }
-}
