+++ src/main/java/wooteco/subway/exception/badRequest/BadRequest.java
@@ -1,0 +1,8 @@
+package wooteco.subway.exception.badRequest;
+
+public class BadRequest extends RuntimeException {
+
+    public BadRequest(String message) {
+        super(message);
+    }
+}
