+++ src/main/java/wooteco/subway/exception/notFound/LineNotFoundException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.exception.notFound;
+
+public class LineNotFoundException extends NotFoundException {
+
+    private static final String MESSAGE = "존재하지 않는 노선입니다.";
+
+    public LineNotFoundException() {
+        super(MESSAGE);
+    }
+}
